*** empty log message ***
git-svn-id: https://svn.nfit.au.dk/svn/beta/trunk@7364 678b6752-a8a0-49ff-a511-7c2a3241fcd4

+++ betarun/C/exit.c
@@ -367,9 +367,9 @@ void BetaError(BetaErr err, Object *theObj)
 	StackEnd = (long *) &theObj; StackEnd++;
 	break;
       }
-      DEBUG_CODE(fprintf(output, "          StackEnd=0x%x,\n", StackEnd));
-      DEBUG_CODE(fprintf(output, "          thePC=0x%x", thePC));
-      DEBUG_CODE(PrintCodeAddress(thePC));
+      DEBUG_CODE(fprintf(output, "          StackEnd=0x%x,\n", (int)StackEnd));
+      DEBUG_CODE(fprintf(output, "          thePC=0x%x", (int)thePC));
+      DEBUG_CODE(PrintCodeAddress((long)thePC));
       DEBUG_CODE(fprintf(output, ")\n"));
 #endif /* intel */
